handling role account

+++ controllers/buyer.js
@@ -10,59 +10,62 @@ const getBuyer = async (req, res) => {
   const offset = (page - 1) * limit;
 
   try {
-    const buyer = await Buyer.findAndCountAll({
-      attributes: ["id", "name", "profile_picture", "phoneNumber"],
-      offset,
-      limit
-    });
+    const accountId = req.role;
 
-    const totalItems = buyer.count;
-    const totalPages = Math.ceil(totalItems / limit);
+    if (accountId === "99") {
+      const buyer = await Buyer.findAndCountAll({
+        attributes: ["id", "name", "profile_picture", "phoneNumber"],
+        offset,
+        limit,
+      });
 
-    if (page > totalPages) {
-      // Jika halaman yang diminta melebihi total halaman yang ada, kirim respons 204 (No Content).
-      return res.status(204).end();
+      const totalItems = buyer.count;
+      const totalPages = Math.ceil(totalItems / limit);
+
+      if (page > totalPages) {
+        // Jika halaman yang diminta melebihi total halaman yang ada, kirim respons 204 (No Content).
+        return res.status(204).end();
+      }
+
+      const response = {
+        message: "Data buyer berhasil diambil!",
+        data: buyer.rows,
+        meta: {
+          totalItems,
+          totalPages,
+          currentPage: page,
+        },
+      };
+
+      if (page > 1) {
+        response.meta.prevPage = page - 1;
+      }
+      if (page < totalPages) {
+        response.meta.nextPage = page + 1;
+      }
+
+      res.status(200).json(response);
+    } else {
+      res.status(403).json({ message: "Tidak diizinkan melihat data buyer!" });
     }
-
-    const response = {
-      message: "Data buyer berhasil diambil!",
-      data: buyer.rows,
-      meta: {
-        totalItems,
-        totalPages,
-        currentPage: page,
-      },
-    };
-
-    if (page > 1) {
-      response.meta.prevPage = page - 1;
-    }
-    if (page < totalPages) {
-      response.meta.nextPage = page + 1;
-    }
-
-    res.status(200).json(response);
   } catch (error) {
     res.status(500).json({ message: error.message });
   }
 };
 
-const getBuyerById = async (req, res) => {
+const getProfile = async (req, res) => {
   try {
+    const accountId = req.user;
+    
     const response = await Buyer.findOne({
-      attributes: ["id", "name", "profile_picture", "phoneNumber"],
+      attributes: ["name", "profile_picture", "phoneNumber"],
       where: {
-        id: req.params.id || null,
+        accountId: accountId,
       },
       include: [
         {
-          model: Unit,
-          attributes: [
-            "uuid", "name", "luas_tanah", "price"
-          ],
-          through: {
-            attributes: []
-          }
+          model: Account,
+          attributes: ['email']
         }
       ]
     });
@@ -73,9 +76,56 @@ const getBuyerById = async (req, res) => {
       });
 
     res.status(200).json({
-      message: "Detail buyer berhasil diambil!",
-      data: response,
+      message: "Data profile berhasil diambil!",
+      data: {
+        name: response.name,
+        email: response.account.email,
+        profilePicture: response.profile_picture,
+        phoneNumber: response.phoneNumber
+      },
     });
+  } catch (error) {
+    res.status(500).json({
+      message: error.message,
+    });
+  }
+}
+
+const getBuyerById = async (req, res) => {
+  try {
+    const accountId = req.role;
+
+    if (accountId === "99") {
+      const response = await Buyer.findOne({
+        attributes: ["id", "name", "profile_picture", "phoneNumber"],
+        where: {
+          id: req.params.id || null,
+        },
+        include: [
+          {
+            model: Unit,
+            attributes: ["uuid", "name", "luas_tanah", "price"],
+            through: {
+              attributes: [],
+            },
+          },
+        ],
+      });
+
+      if (!response)
+        return res.status(404).json({
+          message: "Data tidak ditemukan!",
+        });
+
+      res.status(200).json({
+        message: "Detail buyer berhasil diambil!",
+        data: response,
+      });
+    } else {
+      res
+        .status(403)
+        .json({ message: "Tidak diizinkan melihat detail buyer!" });
+    }
   } catch (error) {
     res.status(500).json({
       message: error.message,
@@ -87,9 +137,17 @@ const createBuyer = async (req, res) => {
   const { accountId, name, phoneNumber, unit_id } = req.body;
 
   try {
+    const idAccount = req.role;
+
     const account = await Account.findByPk(accountId);
 
-    if (account && account.role === "98") {
+    if (!account) {
+      res.status(404).json({ message: "Account tidak ditemukan." });
+    }
+
+    if (idAccount !== "99") {
+      res.status(403).json({ message: "Tidak diizinkan membuat data buyer!" });
+    } else {
       // let profilePictureUrl = req.body.profile_picture;
       let profilePictureUrl = `/public/assets/images/${req.file.filename}`;
 
@@ -113,7 +171,7 @@ const createBuyer = async (req, res) => {
       const buyer = await Buyer.create({
         name: name,
         profile_picture: profilePictureUrl,
-        accountId: account.id,
+        accountId: accountId,
         phoneNumber: phoneNumber,
       });
 
@@ -124,17 +182,18 @@ const createBuyer = async (req, res) => {
           // Menghubungkan unit-unit dengan pembeli
           await buyer.addUnits(units);
         } else {
-          res.status(404).send('Unit tidak ditemukan.');
+          res.status(404).send("Unit tidak ditemukan.");
           return;
         }
       }
 
-      res.status(201).json({ message: "Data Buyer dan unit berhasil ditambahkan dan dihubungkan!" });
-    } else {
-      res.status(403).json({ message: "Tidak diizinkan membuat buyer!" });
+      res.status(201).json({
+        message: "Data Buyer dan unit berhasil ditambahkan dan dihubungkan!",
+      });
     }
   } catch (error) {
     console.log(error);
+
     res.status(500).json({ message: error.message });
   }
 };
@@ -149,12 +208,18 @@ const updateBuyer = async (req, res) => {
   if (!buyer)
     return res.status(404).json({ message: "Data buyer tidak ditemukan!" });
 
+  const account = await Account.findByPk(accountId);
+
+  if (!account) {
+    res.status(404).json({ message: "Account tidak ditemukan." });
+  }
+
   const { accountId, name, phoneNumber } = req.body;
 
   try {
-    const account = await Account.findByPk(accountId);
+    const idAccount = req.role;
 
-    if (account && account.role === "98") {
+    if (idAccount === "99") {
       let profilePictureUrl = `/public/assets/images/${req.file.filename}`;
 
       if (!profilePictureUrl) {
@@ -173,16 +238,19 @@ const updateBuyer = async (req, res) => {
           "https://ui-avatars.com/api/?format=png&name="
         );
       }
-      await Buyer.update({
-        name: name,
-        profile_picture: profilePictureUrl,
-        accountId: account.id,
-        phoneNumber: phoneNumber,
-      },{
-        where: {
-          id: buyer.id,
+      await Buyer.update(
+        {
+          name: name,
+          profile_picture: profilePictureUrl,
+          accountId: accountId,
+          phoneNumber: phoneNumber,
         },
-      });
+        {
+          where: {
+            id: buyer.id,
+          },
+        }
+      );
 
       res.status(201).json({ message: "Buyer berhasil diupdate!" });
     } else {
@@ -205,22 +273,39 @@ const deleteBuyer = async (req, res) => {
       message: "Buyer tidak ditemukan!",
     });
 
-  try {
-    await Buyer.destroy({
-      where: {
-        id: buyer.id,
-      },
-    });
-    res.status(200).json({ message: "Data buyer berhasil dihapus!" });
-  } catch (error) {
-    res.status(400).json({ message: error.message });
+  const account = await Account.findByPk(accountId);
+
+  if (!account) {
+    res.status(404).json({ message: "Account tidak ditemukan." });
   }
-}
+
+  try {
+
+    const accountId = req.role;
+
+    if (accountId === "99") {
+
+      await Buyer.destroy({
+        where: {
+          id: buyer.id,
+        },
+      });
+      res.status(200).json({ message: "Data buyer berhasil dihapus!" });
+    } else {
+      res
+        .status(403)
+        .json({ message: "Tidak diizinkan menghapus data buyer!" });
+    }
+  } catch (error) {
+    res.status(500).json({ message: error.message });
+  }
+};
 
 module.exports = {
   createBuyer,
   getBuyer,
+  getProfile,
   getBuyerById,
   updateBuyer,
-  deleteBuyer
+  deleteBuyer,
 };
